20581 improve external type resolvers
extract mapping layer functionality to separate class
